Made Student a subclass of User
For separation of concerns
